change to add_theme_page params
WP_DEBUG was throwing notices for the add_theme_page function on line 118. I changed the user capabilities argument and added quotes to the callback function (without the quotes WP was seeing that argument as a global). I'm new to pull requests so let me know if these are correct changes!